CBD-6212, CBD-6213: Update Black Duck scripts, fix Trinity scans
Black Duck are migrating their infrastructure from the old Synopsis
domain. We need to ensure we are pointing at the new URL to avoid
interruption. This change also uses `logging.level.detect` in place
of the now deprecated `logging.level.com.synopsys.integration`,
upgrades the default detect.jar we're using to current while
ensuring we can still download the working 9.x versions if required
(the version we're currently using will stop working on the 14th)
and updates some outdated synopsys URLs.

In addition to this, Trinity scans are currently failing as a result
of a recent upgrade to a version which includes dart, rust and swift
files automatically detected by black duck, but which we do not
ingest, this change ensures they are purged prior to the scan.

Change-Id: Iac8958cf4bd8ba68f8ba4b6f459b8256e11d7d2e
Reviewed-on: https://review.couchbase.org/c/build-tools/+/223140
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Blair Watt <blair.watt@couchbase.com>