Black Duck: Update JVM SDK configs for new tagging scheme
Motivation
----------
Starting with version 3.9.0, the couchbase-jvm-clients monorepo
now uses a mono-versioning scheme. Instead of tagging the SDKs
individually (as `java-client-a.b.c`, `scala-client-x.y.z`, etc.)
the SDKs share a common tag that is just the version (like `3.9.0`).

Modifications
-------------
Remove the logic for scanning Java SDK 2.x since we don't
plan to release any more versions from that line.

In each SDK's `get_source.sh`, set TAG to just the version
unless the version is below the jump to 3.9.0.

Use glob match syntax to check the version, because it's
simpler than regex and sufficient for the version prefix check.

Update scan config because Kotlin SDK 1.6.0
and Scala SDK 1.9.0 were both released as 3.9.0.

Change-Id: I7f63006c5a92174012ada8b6abf66d4956e6e8b3
Reviewed-on: https://review.couchbase.org/c/build-tools/+/232005
Tested-by: David Nault <david.nault@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>