couchbase-java-sdk patch
The main fix is to add compile so that all modules are
built before dependency resolution is attempted.  This
allows the script to work with SDKs that aren't yet
published to Maven Central.

Also switching to ./mvnw as it may be required to
use recent Maven plugins.

And the script's current error flow is confusing:
it tries dependency resolution, and if that fails
then it installs all dependencies (basically what
the Makefile does) and then.. stops.  Maybe for
the next run to pickup?

Replaced that with just installing the dependencies
upfront.