Fix up Maven-related SDK Black Duck scans
For various reasons, the products from couchbase-jvm-clients have
incomplete inter-module dependency tracking, meaning a simple `mvn
dependency:resolve` at the top level isn't sufficient to populate the
local Maven repository for Black Duck. Additionally, not all built
artifacts are available in Maven repos: for `master` scans, no SNAPSHOT
jars are; and for all versions, the various test jars are not. So in the
end, it seems like the only way to get everything available locally is
to build the whole project with a top-level `make install`.

One positive is that `make install` will use the `./mvnw` wrapper,
meaning we no longer need to install Maven with `cbdep` and should
always run with the correct Maven version. We also copy `mvnw` into
every subdirectory with a `pom.xml` so that Black Duck will use the same
wrapper for each submodule.

Also increase the scanning intervals for tagged releases, since they
should be roughly constant anyway.

Change-Id: I5c5a79f755b98f027853aef78d455d1b348d989d
Reviewed-on: https://review.couchbase.org/c/build-tools/+/232485
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>