CBD-6501: Fix black duck new version detection
The bd new version detection script is currently hitting a fatal
error thanks to a tag which cannot be decoded. This change ensures
we use git for-each-ref instead of GitPython's tag iterator to avoid
UnicodeDecodeError on repos with malformed tag names. It also fetches
all tags after clone to ensure commit objects are available, and
silently skips stale refs that no longer resolve.

Change-Id: I9ea54db1c45dc9edc030efad7fff9792f6d6d252
Reviewed-on: https://review.couchbase.org/c/build-tools/+/236087
Reviewed-by: Ming Ho <ming.ho@couchbase.com>
Tested-by: Blair Watt <blair.watt@couchbase.com>